Create controller & service
