+++ app/Http/Controllers/common/ProductController.php
@@ -2,9 +2,6 @@
 
 namespace App\Http\Controllers\common;
 
-use App\Http\Requests\CartRequest;
-use App\Interfaces\SessionKeyInterface;
-use App\Models\Cart;
 use App\Models\Product;
 use App\Models\ErrorLog;
 use App\Models\Category;
@@ -15,9 +12,8 @@ use App\Utilities\GoogleDriveUtility;
 use App\Utilities\ProductsUtility;
 use App\Interfaces\StatusInterface;
 use App\Http\Controllers\Controller;
-use Illuminate\Support\Facades\DB;
 
-class ProductController extends Controller implements StatusInterface, SessionKeyInterface
+class ProductController extends Controller implements StatusInterface
 {
     private $googleDriveUtility;
     private $productUtility;
@@ -93,52 +89,5 @@ class ProductController extends Controller implements StatusInterface, SessionKe
             'message' => 'Data sorted!',
             'data' => $products,
         ], Response::HTTP_OK);
-    }
-
-    public function addToCart(CartRequest $cartRequest)
-    {
-        $item = $cartRequest->validated();
-        /**
-         * @var \App\Models\User $user
-         */
-        $user = session(self::SESSION_IDENTITY);
-
-        try {
-            DB::beginTransaction();
-            
-            $product = Product::where('product_serial_code', $item['product_serial'])->first();
-            
-            if (!$product) {
-                throw new \Exception('Invalid operation.');
-            }
-
-            $cart = new Cart();
-            $cart->user_id = $user->id;
-            $cart->product_id = $product->id;
-            $cart->quantity = $item['quantity'];
-            $cart->save();
-            
-            DB::commit();
-        } catch (\Exception $e) {
-            DB::rollBack();
-
-            $errorLog = new ErrorLog();
-            $errorLog->error = $e->getMessage();
-            $errorLog->save();
-
-            $response = [
-                'status' => self::STATUS_ERROR,
-                'message' => 'Invalid operation.',
-            ];
-
-            return back()->withInput()->with($response);
-        }
-
-        $response = [
-            'status' => self::STATUS_SUCCESS,
-            'message' => 'Product added to cart.',
-        ];
-
-        return back()->with($response);
     }
 }
